+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix #4534: Java Generator CLI default handling of skipGeneratedAnnotations
 * Fix #4535: The shell command string will now have single quotes sanitized
 * Fix #4547: preventing timing issues with leader election cancel
+* Fix #4564: IOException when reading response body is swallowed by OkHttp based implementation 
 
 #### Improvements